requirements.txt
Django==2.1.3
djangorestframework==3.9.0
django-environ==0.4.5
factory_boy==2.11.1
psycopg2-binary==2.7.6.1
coverage==4.5.2
